scripts/requirements.txt
PyYAML
pysolr==3.6
ontobio==2.8.4
click
pyparsing==2.4.7
antlr4-python3-runtime==4.9.3
gocam
